feat(llm): upgrade default Claude model to Sonnet 4.0
* Updated default Claude model from Claude 3.7 Sonnet to Claude Sonnet 4.0 across multiple configuration files and code references.
* Replaced legacy model constants with new model version in configuration, documentation, and initialization settings.
* Updated README and development documentation to reflect the new recommended model for standard requests.

The changes include updates to configuration files, CLI initialization, documentation, and model-related code to standardize the use of Claude Sonnet 4.0 as the default model.
